requirements.txt
numpy>=1.21.0
Pillow>=8.0.0 